requirements.txt
Django==5.1.6
django_ratelimit==4.1.0
djangorestframework==3.15.2
python-dotenv==1.0.1
web3==7.8.0
colorlog==6.9.0
